+++ src/Runtime/IO.c
@@ -2,7 +2,7 @@
  *                             IO                                 *
  *----------------------------------------------------------------*/
 #include <stdlib.h>
-#include <sys/param.h> 
+#include <sys/param.h>
 #include <sys/stat.h>
 #include <sys/poll.h>
 #include <dirent.h>
@@ -24,11 +24,11 @@
 #include "Math.h"
 #include "Runtime.h"
 
-uintptr_t 
+uintptr_t
 openInStream(String path, uintptr_t exn)                    /* SML Basis */
-{              
+{
   FILE *fileDesc;
-  if ((fileDesc = fopen(&(path->data), "r")) == NULL) 
+  if ((fileDesc = fopen(&(path->data), "r")) == NULL)
     {
       raise_exn(exn);
     }
@@ -36,11 +36,11 @@ openInStream(String path, uintptr_t exn)                    /* SML Basis */
   return (uintptr_t)(tag_scalar(fileDesc));
 }
 
-uintptr_t 
+uintptr_t
 openOutStream(String path, uintptr_t exn)                   /* SML Basis */
 {
   FILE *fileDesc;
-  if ((fileDesc = fopen(&(path->data), "w")) == NULL) 
+  if ((fileDesc = fopen(&(path->data), "w")) == NULL)
     {
       raise_exn(exn);
     }
@@ -48,11 +48,11 @@ openOutStream(String path, uintptr_t exn)                   /* SML Basis */
   return (uintptr_t)(tag_scalar(fileDesc));
 }
 
-uintptr_t 
+uintptr_t
 openAppendStream(String path, uintptr_t exn)                /* SML Basis */
 {
   FILE *fileDesc;
-  if ((fileDesc = fopen(&(path->data), "a")) == NULL) 
+  if ((fileDesc = fopen(&(path->data), "a")) == NULL)
     {
       raise_exn(exn);
     }
@@ -60,11 +60,11 @@ openAppendStream(String path, uintptr_t exn)                /* SML Basis */
   return (uintptr_t)(tag_scalar(fileDesc));
 }
 
-uintptr_t 
+uintptr_t
 openInBinStream(String path, uintptr_t exn)                 /* SML Basis */
 {
   FILE *fileDesc;
-  if ((fileDesc = fopen(&(path->data), "rb")) == NULL) 
+  if ((fileDesc = fopen(&(path->data), "rb")) == NULL)
     {
       raise_exn(exn);
     }
@@ -72,23 +72,23 @@ openInBinStream(String path, uintptr_t exn)                 /* SML Basis */
   return (uintptr_t)(tag_scalar(fileDesc));
 }
 
-uintptr_t 
+uintptr_t
 openOutBinStream(String path, uintptr_t exn)                /* SML Basis */
 {
   FILE *fileDesc;
-  if ((fileDesc = fopen(&(path->data), "wb")) == NULL) 
+  if ((fileDesc = fopen(&(path->data), "wb")) == NULL)
     {
       raise_exn(exn);
-    }  
+    }
   check_tag_scalar(fileDesc);
   return (uintptr_t)(tag_scalar(fileDesc));
 }
 
-uintptr_t 
+uintptr_t
 openAppendBinStream(String path, uintptr_t exn)             /* SML Basis */
 {
   FILE *fileDesc;
-  if ((fileDesc = fopen(&(path->data), "ab")) == NULL) 
+  if ((fileDesc = fopen(&(path->data), "ab")) == NULL)
     {
       raise_exn(exn);
     }
@@ -114,11 +114,11 @@ input1Stream(uintptr_t is1)
 
 
 // inputStream:
-//   Reads n characters from input, n<=64. If EOF is read, 
+//   Reads n characters from input, n<=64. If EOF is read,
 //   then a string less than n is returned.
 
 String
-REG_POLY_FUN_HDR(inputStream, Region rd, uintptr_t is1, size_t n) 
+REG_POLY_FUN_HDR(inputStream, Region rd, uintptr_t is1, size_t n)
 {
   char buf[100];
   size_t i;
@@ -140,9 +140,9 @@ REG_POLY_FUN_HDR(inputStream, Region rd, uintptr_t is1, size_t n)
       resetRegion(rd);
     }
 
-  //  i = fread(buf,1,n,is); 
+  //  i = fread(buf,1,n,is);
   //  return REG_POLY_CALL(convertBinStringToML, rd, i, buf);
-    
+
   terminal = isatty(fileno(is));
   for ( i = 0; i < n && ((ch = fgetc(is)) != EOF); i++ )
     {
@@ -154,7 +154,7 @@ REG_POLY_FUN_HDR(inputStream, Region rd, uintptr_t is1, size_t n)
 }
 
 size_t
-lookaheadStream(uintptr_t is1) 
+lookaheadStream(uintptr_t is1)
 {
   int ch;
   FILE *is;
@@ -168,8 +168,8 @@ lookaheadStream(uintptr_t is1)
   return convertIntToML(ch);
 }
 
-void 
-closeStream(uintptr_t stream) 
+void
+closeStream(uintptr_t stream)
 {
   stream = untag_scalar(stream);
   fclose((FILE *) stream);
@@ -177,8 +177,8 @@ closeStream(uintptr_t stream)
 }
 
 /*
-int 
-endOfStream(FILE *stream) 
+int
+endOfStream(FILE *stream)
 {
   int ch;
 
@@ -193,8 +193,8 @@ endOfStream(FILE *stream)
 }
 */
 
-size_t 
-outputStream(uintptr_t os1, String s, uintptr_t exn) 
+size_t
+outputStream(uintptr_t os1, String s, uintptr_t exn)
 {
   FILE *os = (FILE *)untag_scalar(os1);
   if ( fputs(&(s->data), os) == EOF )
@@ -205,79 +205,79 @@ outputStream(uintptr_t os1, String s, uintptr_t exn)
   return mlUNIT;
 }
 
-void 
-flushStream(uintptr_t stream) 
+void
+flushStream(uintptr_t stream)
 {
   stream = untag_scalar(stream);
   fflush((FILE *) stream); /* What about error. */
 }
 
-size_t 
-stdInStream(uintptr_t dummy) 
+size_t
+stdInStream(uintptr_t dummy)
 {
   check_tag_scalar(stdin);
   return (size_t)tag_scalar(stdin);
 }
 
-size_t 
-stdOutStream(size_t dummy) 
+size_t
+stdOutStream(size_t dummy)
 {
   check_tag_scalar(stdout);
   return (size_t)tag_scalar(stdout);
 }
 
-size_t 
-stdErrStream(uintptr_t dummy) 
+size_t
+stdErrStream(uintptr_t dummy)
 {
   check_tag_scalar(stderr);
   return (size_t)tag_scalar(stderr);
 }
 
-void 
-sml_chdir(String dirname, long exn)              /* SML Basis */
+void
+sml_chdir(String dirname, uintptr_t exn)              /* SML Basis */
 {
-  if ( chdir(&(dirname->data)) != 0 ) 
+  if ( chdir(&(dirname->data)) != 0 )
     {
       raise_exn(exn);
     }
   return;
 }
 
-void 
-sml_remove(String name, int exn)                /* SML Basis */
+void
+sml_remove(String name, uintptr_t exn)                /* SML Basis */
 {
   int ret;
   ret = unlink(&(name->data));
-  if ( ret != 0 ) 
+  if ( ret != 0 )
     {
       raise_exn(exn);
     }
   return;
 }
 
-void 
-sml_rename(String oldname, String newname, int exn)    /* SML Basis */
+void
+sml_rename(String oldname, String newname, uintptr_t exn)    /* SML Basis */
 {
-  if ( rename(&(oldname->data), &(newname->data)) != 0 ) 
+  if ( rename(&(oldname->data), &(newname->data)) != 0 )
     {
       raise_exn(exn);
     }
   return;
 }
 
-int 
-sml_access(String path, int permarg, int exn)               /* ML */
+size_t
+sml_access(String path, size_t permarg, uintptr_t exn)               /* ML */
 {
   long perms;
   long perm = convertIntToC(permarg);
   perms  = ((0x1 & perm) ? R_OK : 0);
   perms |= ((0x2 & perm) ? W_OK : 0);
   perms |= ((0x4 & perm) ? X_OK : 0);
-  if (perms == 0) 
+  if (perms == 0)
     {
       perms = F_OK;
     }
-  if (access(&(path->data), perms) == 0) 
+  if (access(&(path->data), perms) == 0)
     {
       return mlTRUE;
     }
@@ -285,38 +285,38 @@ sml_access(String path, int permarg, int exn)               /* ML */
 }
 
 String
-REG_POLY_FUN_HDR(sml_getdir, Region rAddr, int exn)                 	 /* SML Basis */
+REG_POLY_FUN_HDR(sml_getdir, Region rAddr, uintptr_t exn)                 	 /* SML Basis */
 {
  char directory[MAXPATHLEN];
  char *res;
  errno = 0;
  res = getcwd(directory, MAXPATHLEN);
- if ( res == NULL ) 
+ if ( res == NULL )
    {
-     raise_exn(exn); 
+     raise_exn(exn);
    }
  return REG_POLY_CALL(convertStringToML, rAddr, directory);
 }
 
-int 
-sml_isdir(String path, int exn)             /* SML Basis */
+size_t
+sml_isdir(String path, uintptr_t exn)             /* SML Basis */
 {
   struct stat buf;
-  if ( stat(&(path->data), &buf) == -1 ) 
+  if ( stat(&(path->data), &buf) == -1 )
     {
       raise_exn(exn);
     }
-  if (S_ISDIR(buf.st_mode)) 
+  if (S_ISDIR(buf.st_mode))
     {
       return mlTRUE;
     }
   return mlFALSE;
 }
 
-void 
-sml_mkdir(String path, int exn)                        /* SML Basis */
+void
+sml_mkdir(String path, uintptr_t exn)                        /* SML Basis */
 {
-  if ( mkdir(&(path->data), 0777) == -1 ) 
+  if ( mkdir(&(path->data), 0777) == -1 )
     {
       raise_exn(exn);
     }
@@ -324,11 +324,11 @@ sml_mkdir(String path, int exn)                        /* SML Basis */
 }
 
 
-int 
-sml_modtime(uintptr_t vAddr, String path, int exn)             /* SML Basis */
+uintptr_t
+sml_modtime(uintptr_t vAddr, String path, uintptr_t exn)             /* SML Basis */
 {
   struct stat buf;
-  if ( stat(&(path->data), &buf) == -1 ) 
+  if ( stat(&(path->data), &buf) == -1 )
     {
       raise_exn(exn);
     }
@@ -337,61 +337,61 @@ sml_modtime(uintptr_t vAddr, String path, int exn)             /* SML Basis */
   return vAddr;
 }
 
-void 
-sml_rmdir(String path, int exn)              /* SML Basis */
+void
+sml_rmdir(String path, uintptr_t exn)              /* SML Basis */
 {
-  if ( rmdir(&(path->data)) == -1 ) 
+  if ( rmdir(&(path->data)) == -1 )
     {
       raise_exn(exn);
     }
   return;
 }
 
-void 
-sml_settime(String path, uintptr_t time, int exn)     /* SML Basis */
+void
+sml_settime(String path, uintptr_t time, uintptr_t exn)     /* SML Basis */
 {
   struct utimbuf tbuf;
   tbuf.actime = tbuf.modtime = (long)(get_d(time));
-  if ( utime(&(path->data), &tbuf) == -1 ) 
+  if ( utime(&(path->data), &tbuf) == -1 )
     {
       raise_exn(exn);
     }
   return;
 }
 
-int 
-sml_filesize(String path, int exn)              /* SML Basis */
+size_t
+sml_filesize(String path, uintptr_t exn)              /* SML Basis */
 {
   struct stat buf;
-  if ( stat(&(path->data), &buf) == -1 ) 
+  if ( stat(&(path->data), &buf) == -1 )
     {
       raise_exn(exn);
     }
   return (convertIntToML(buf.st_size));
 }
 
-uintptr_t 
-sml_opendir(String path, int exn)           /* SML Basis */
+uintptr_t
+sml_opendir(String path, uintptr_t exn)           /* SML Basis */
 {
-  DIR * dstr;    
+  DIR * dstr;
   dstr = opendir(&(path->data));
-  if ( dstr == NULL ) 
+  if ( dstr == NULL )
     {
       raise_exn(exn);
     }
   check_tag_scalar(dstr);
-  return (uintptr_t)tag_scalar(dstr); 
+  return (uintptr_t)tag_scalar(dstr);
 }
 
 String
-REG_POLY_FUN_HDR(sml_readdir, Region rAddr, uintptr_t v, int exn)    /* SML Basis */
+REG_POLY_FUN_HDR(sml_readdir, Region rAddr, uintptr_t v, uintptr_t exn)    /* SML Basis */
 {
   struct dirent *direntry;
   String res;
   DIR * dir_ptr;
   dir_ptr = (DIR *)untag_scalar(v);
   direntry = readdir(dir_ptr);
-  if (direntry == NULL) 
+  if (direntry == NULL)
     {
       raise_exn(exn);
       return NULL;
@@ -403,7 +403,7 @@ REG_POLY_FUN_HDR(sml_readdir, Region rAddr, uintptr_t v, int exn)    /* SML Basi
   return res;
 }
 
-void 
+void
 sml_rewinddir(uintptr_t v)            /* SML Basis */
 {
   DIR *dir_ptr;
@@ -413,20 +413,20 @@ sml_rewinddir(uintptr_t v)            /* SML Basis */
   return;
 }
 
-void 
-sml_closedir(uintptr_t v, int exn)            /* SML Basis */
+void
+sml_closedir(uintptr_t v, uintptr_t exn)            /* SML Basis */
 {
   DIR *dir_ptr;
 
   dir_ptr = (DIR *)untag_scalar(v);
-  if (closedir(dir_ptr) == -1) 
-    { 
+  if (closedir(dir_ptr) == -1)
+    {
       raise_exn(exn);
     }
   return;
 }
 
-size_t 
+size_t
 sml_errno(void)             /* SML Basis */
 {
   return convertIntToML((size_t) errno);                 // not thread-safe!!
@@ -434,7 +434,7 @@ sml_errno(void)             /* SML Basis */
 
 // FIXME
 String
-REG_POLY_FUN_HDR(sml_errormsg, Region rAddr, int errnum)    /* SML Basis */
+REG_POLY_FUN_HDR(sml_errormsg, Region rAddr, size_t errnum)    /* SML Basis */
 {
   char *res;
   res = strerror(convertIntToC(errnum));
@@ -443,41 +443,41 @@ REG_POLY_FUN_HDR(sml_errormsg, Region rAddr, int errnum)    /* SML Basis */
   return REG_POLY_CALL(convertStringToML, rAddr, res);
 }
 
-int 
-sml_islink(String path, int exn)              /* SML Basis */
+size_t
+sml_islink(String path, uintptr_t exn)              /* SML Basis */
 {
   struct stat buf;
-  if (lstat(&(path->data), &buf) == -1) 
+  if (lstat(&(path->data), &buf) == -1)
     {
       raise_exn(exn);
     }
-  if (S_ISLNK(buf.st_mode)) 
-    { 
+  if (S_ISLNK(buf.st_mode))
+    {
       return mlTRUE;
     }
   return mlFALSE;
 }
 
-int 
-sml_isreg(int fd, int exn)              /* SML Basis */
+size_t
+sml_isreg(size_t fd, uintptr_t exn)              /* SML Basis */
 {
   struct stat buf;
-  if (fstat(convertIntToC(fd), &buf) == -1) 
+  if (fstat(convertIntToC(fd), &buf) == -1)
     {
       raise_exn(exn);
     }
-  if (S_ISREG(buf.st_mode)) 
-    { 
+  if (S_ISREG(buf.st_mode))
+    {
       return mlTRUE;
     }
   return mlFALSE;
 }
 
-int 
-sml_filesizefd(int fd, int exn)              /* SML Basis */
+size_t
+sml_filesizefd(size_t fd, uintptr_t exn)              /* SML Basis */
 {
   struct stat buf;
-  if (fstat(convertIntToC(fd), &buf) == -1) 
+  if (fstat(convertIntToC(fd), &buf) == -1)
     {
       raise_exn(exn);
     }
@@ -485,14 +485,14 @@ sml_filesizefd(int fd, int exn)              /* SML Basis */
 }
 
 String
-REG_POLY_FUN_HDR(sml_readlink, Region rAddr, String path, int exn)    /* SML Basis */
+REG_POLY_FUN_HDR(sml_readlink, Region rAddr, String path, uintptr_t exn)    /* SML Basis */
 {
   char buffer[MAXPATHLEN];
   long result;
   result = readlink(&(path->data), buffer, MAXPATHLEN);
-  if (result == -1 || result >= MAXPATHLEN) 
+  if (result == -1 || result >= MAXPATHLEN)
     {
-      raise_exn(exn); 
+      raise_exn(exn);
     }
   buffer[result] = '\0';
   return REG_POLY_CALL(convertStringToML, rAddr, buffer);
@@ -501,12 +501,12 @@ REG_POLY_FUN_HDR(sml_readlink, Region rAddr, String path, int exn)    /* SML Bas
 extern char *realpath();
 
 String
-REG_POLY_FUN_HDR(sml_realpath, Region rAddr, String path, int exn)  /* SML Basis */
+REG_POLY_FUN_HDR(sml_realpath, Region rAddr, String path, uintptr_t exn)  /* SML Basis */
 {
   char buffer[MAXPATHLEN];
   char *result;
   result = realpath(&(path->data), buffer);
-  if (result == NULL) 
+  if (result == NULL)
     {
       raise_exn(exn);
       return NULL;
@@ -514,11 +514,11 @@ REG_POLY_FUN_HDR(sml_realpath, Region rAddr, String path, int exn)  /* SML Basis
   return REG_POLY_CALL(convertStringToML, rAddr, result);
 }
 
-uintptr_t 
-sml_devinode(uintptr_t vAddr, String path, int exn)             /* SML Basis */
+uintptr_t
+sml_devinode(uintptr_t vAddr, String path, uintptr_t exn)             /* SML Basis */
 {
   struct stat buf;
-  if (stat(&(path->data), &buf) == -1) 
+  if (stat(&(path->data), &buf) == -1)
     {
       raise_exn(exn);
     }
@@ -529,12 +529,12 @@ sml_devinode(uintptr_t vAddr, String path, int exn)             /* SML Basis */
   return vAddr;
 }
 
-int 
-sml_system(String cmd, int exn)         /* SML Basis */
+size_t
+sml_system(String cmd, uintptr_t exn)         /* SML Basis */
 {
   int res;
   res = system(&(cmd->data));
-  if (res != 0) 
+  if (res != 0)
     {
       res = -1;
     }
@@ -542,22 +542,22 @@ sml_system(String cmd, int exn)         /* SML Basis */
 }
 
 String
-REG_POLY_FUN_HDR(sml_getenv, Region rAddr, String var, int exn)  /* SML Basis */
+REG_POLY_FUN_HDR(sml_getenv, Region rAddr, String var, uintptr_t exn)  /* SML Basis */
 {
   char *res;
   res = (char *)(getenv(&(var->data)));
-  if (res == NULL) 
+  if (res == NULL)
     {
       raise_exn(exn);
     }
   return REG_POLY_CALL(convertStringToML, rAddr, res);
 }
 
-size_t 
-outputBinStream(uintptr_t os1, String s, uintptr_t exn) 
+size_t
+outputBinStream(uintptr_t os1, String s, uintptr_t exn)
 { long strsize;
   FILE *os = (FILE *) os1;
-  strsize = sizeStringDefine(s); 
+  strsize = sizeStringDefine(s);
   os = (FILE *)untag_scalar(os);
   if ( fwrite(&(s->data), 1, strsize, os) != strsize )
     {
@@ -568,14 +568,14 @@ outputBinStream(uintptr_t os1, String s, uintptr_t exn)
 }
 
 uintptr_t
-sml_microsleep(uintptr_t pair, int s, int u)
+sml_microsleep(uintptr_t pair, size_t s, size_t u)
 {
-  int r;
+  size_t r;
   struct timespec req, rem;
   mkTagPairML(pair);
   u = convertIntToC(u);
   s = convertIntToC(s);
-  while (u > 1000000) 
+  while (u > 1000000)
   {
     s++;
     u -= 1000000;
@@ -589,11 +589,10 @@ sml_microsleep(uintptr_t pair, int s, int u)
   return pair;
 }
 
-int
-sml_poll(int time)
+size_t
+sml_poll(size_t time)
 {
-  int r;
+  size_t r;
   r = poll(0,0,time);
   return r;
 }
-
